Making front and backend https using self signed cert
Making front and backend https.
Notes: The localhost.conf are the values used for creating the self signed cert. Also, changed backend controllers to https and the "theAllowedOrigins" value to https, but did not change the frontend request URLs https. Not sure why. I would think we would need to do that. Also, the .env file is what's used to set the values which point to the location of the .crt and .key. Non of this works without the change to the package.json which is different for Windows vs Linus & Mac.